requirements.txt
pyrogram==2.0.106
tgcalls==3.0.0.dev6
motor==3.3.1
pymongo==4.5.0
python-dotenv==1.0.1
yt-dlp==2024.12.13
tgcrypto==1.2.5
aiohttp==3.10.11
ffmpeg-python==0.2.0